requirements.txt
# SketchMind Requirements
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=9.0.0
networkx>=3.1
requests>=2.28.0

# Optional: Only needed for local model loading (not required for OpenRouter)
# Uncomment if you plan to use local Llama models instead of OpenRouter
# torch>=2.1.0
# transformers>=4.39.0
# accelerate
# torchvision
# tqdm>=4.64.0
# joblib>=1.3.0
